+++ Source/Sitecore.FakeDb/Security/Accounts/FakeMembershipUser.cs
@@ -1,0 +1,11 @@
+﻿namespace Sitecore.FakeDb.Security.Accounts
+{
+  using System.Web.Security;
+
+  public class FakeMembershipUser : MembershipUser
+  {
+    public FakeMembershipUser()
+    {
+    }
+  }
+}